HV-381: Add failFast option to ValidatorContextImpl, remove HibernateValidatorContextImpl
